feat(app-update): a real updates page instead of a pop-up and a link to GitHub
The update flow was split across three places that each did part of the job
badly: a sidebar badge that opened a browser, a dialog that opened itself over
whatever the operator was doing, and a settings panel that only held the
download folder. The changelog in that dialog was the raw release body — a wall
of markdown with `##` headings and `**bold**` left in.

Settings -> Updates is now the whole flow, and the only place it lives:

- current version, the new one beside it, and a Check button whose icon spins
  while it checks — without that, a check that finds nothing looked like a dead
  click
- what the new version brings, rendered with the same components the release
  notes history uses: grouped into New / Fixed / Changed, each entry with its
  scope, instead of markdown
- Download, then a progress bar with percent and MB, then Install

The self-opening dialog is gone.

The sidebar badge now opens that page rather than GitHub, and keeps showing
once the installer is downloaded but not yet run — the moment the reminder
matters most, and previously the moment it disappeared. It mirrors the download
progress too, so switching away from the page does not hide that something is in
flight, and a downloaded update outranks "dismiss": the operator asked for it,
so it stays until it is applied.

Also removes `VersionDisplay`, which nothing rendered.
