+++ app/apps/client/src/features/app-update/index.ts
@@ -1,8 +1,6 @@
 export { AboutSection } from './components/AboutSection'
-export { UpdateAvailableModal } from './components/UpdateAvailableModal'
 export { UpdateNotification } from './components/UpdateNotification'
-export { UpdateSettingsPanel } from './components/UpdateSettingsPanel'
-export { VersionDisplay } from './components/VersionDisplay'
+export { UpdatePanel } from './components/UpdatePanel'
 export { useAppUpdate } from './hooks/useAppUpdate'
 export { useUpdateDownload } from './hooks/useUpdateDownload'
 export type { GithubRelease, UpdateInfo } from './services/versionService'
