+++ app/apps/client/src/i18n/locales/en/sidebar.json
@@ -28,6 +28,11 @@
     "clickToDownload": "Click to download the latest version",
     "download": "Download Update",
     "downloading": "Opening...",
+    "readyToInstall": "v{{version}} ready to install",
+    "clickToOpenUpdates": "Open the updates page",
+    "clickToInstall": "Downloaded — install it now",
+    "viewUpdate": "See what's new",
+    "install": "Install now",
     "dismiss": "Dismiss",
     "dev": "Dev",
     "devInstanceHint": "Running a dev instance — update checks are disabled"
