+++ app/apps/client/src/i18n/locales/ro/sidebar.json
@@ -28,6 +28,11 @@
     "clickToDownload": "Apasă pentru a descărca ultima versiune",
     "download": "Descarcă Actualizarea",
     "downloading": "Se deschide...",
+    "readyToInstall": "v{{version}} gata de instalare",
+    "clickToOpenUpdates": "Deschide pagina de actualizari",
+    "clickToInstall": "Descarcata - instaleaz-o acum",
+    "viewUpdate": "Vezi ce e nou",
+    "install": "Instaleaza acum",
     "dismiss": "Închide",
     "dev": "Dev",
     "devInstanceHint": "Rulează o instanță de dezvoltare — verificarea actualizărilor este dezactivată"
