+++ app/apps/client/src/routes/__root.tsx
@@ -9,7 +9,6 @@ import { useEffect } from 'react'
 
 import { isTauri } from '~/config'
 import { MobileConnectionGuard } from '~/features/api-url-config'
-import { UpdateAvailableModal } from '~/features/app-update'
 import { LoginGate } from '~/features/auth'
 import {
   GlobalAppShortcutManager,
@@ -171,9 +170,6 @@ function MainLayout() {
                                   onto Marcaje / Programe. Global because the
                                   drag crosses subtrees. */}
                               <SongDragGhost />
-                              {/* Opens by itself when a newer version is
-                                  published; postponing is per-version. */}
-                              <UpdateAvailableModal />
                               <AutoOpenScreens />
                               <AutoReopenScreensOnPresentation />
                               <AutoCloseScreensOnHide />
